Move tile layer related files under layer.tiles
