+++ src/androidTest/java/de/blau/android/layer/LayerDialogTest.java
@@ -46,6 +46,7 @@ import de.blau.android.R;
 import de.blau.android.TestUtils;
 import de.blau.android.gpx.GpxTest;
 import de.blau.android.layer.data.MapOverlay;
+import de.blau.android.layer.tiles.MapTilesLayer;
 import de.blau.android.osm.BoundingBox;
 import de.blau.android.osm.Node;
 import de.blau.android.osm.StorageDelegator;
@@ -56,7 +57,6 @@ import de.blau.android.resources.TileLayerDatabase;
 import de.blau.android.resources.TileLayerSource;
 import de.blau.android.tasks.TaskStorage;
 import de.blau.android.tasks.Todo;
-import de.blau.android.views.layers.MapTilesLayer;
 import okhttp3.mockwebserver.MockWebServer;
 
 @RunWith(AndroidJUnit4.class)
