+++ src/main/java/de/blau/android/imageryoffset/ImageryAlignmentActionModeCallback.java
@@ -58,6 +58,7 @@ import de.blau.android.contract.Schemes;
 import de.blau.android.dialogs.Progress;
 import de.blau.android.easyedit.EasyEditActionModeCallback;
 import de.blau.android.imageryoffset.ImageryOffset.DeprecationNote;
+import de.blau.android.layer.tiles.MapTilesLayer;
 import de.blau.android.osm.Server;
 import de.blau.android.osm.UserDetails;
 import de.blau.android.osm.ViewBox;
@@ -72,7 +73,6 @@ import de.blau.android.util.SavingHelper;
 import de.blau.android.util.ScreenMessage;
 import de.blau.android.util.SerializableState;
 import de.blau.android.util.ThemeUtils;
-import de.blau.android.views.layers.MapTilesLayer;
 import okhttp3.Call;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
