+++ src/main/java/de/blau/android/layer/tiles/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Classes for map layers for example using map tiles
  */
-package de.blau.android.views.layers;
+package de.blau.android.layer.tiles;
