+++ src/main/java/de/blau/android/resources/TileLayerSource.java
@@ -60,6 +60,7 @@ import de.blau.android.contract.Files;
 import de.blau.android.contract.MimeTypes;
 import de.blau.android.contract.Paths;
 import de.blau.android.imageryoffset.Offset;
+import de.blau.android.layer.tiles.MapTilesLayer;
 import de.blau.android.osm.BoundingBox;
 import de.blau.android.osm.Server;
 import de.blau.android.osm.ViewBox;
@@ -76,7 +77,6 @@ import de.blau.android.util.FileUtil;
 import de.blau.android.util.GeoMath;
 import de.blau.android.util.Util;
 import de.blau.android.util.Version;
-import de.blau.android.views.layers.MapTilesLayer;
 import okhttp3.Call;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
