+++ src/main/java/de/blau/android/services/util/MapTileDownloader.java
@@ -36,12 +36,12 @@ import de.blau.android.App;
 import de.blau.android.R;
 import de.blau.android.contract.MimeTypes;
 import de.blau.android.exception.InvalidTileException;
+import de.blau.android.layer.tiles.util.MapTileProvider;
+import de.blau.android.layer.tiles.util.MapTileProviderCallback;
 import de.blau.android.resources.TileLayerSource;
 import de.blau.android.resources.TileLayerSource.Header;
 import de.blau.android.util.NetworkStatus;
 import de.blau.android.util.OkHttpFileChannel;
-import de.blau.android.views.util.MapTileProvider;
-import de.blau.android.views.util.MapTileProviderCallback;
 import okhttp3.Call;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
