+++ src/test/java/de/blau/android/services/util/MBTMapTileFilesystemProviderTest.java
@@ -21,12 +21,12 @@ import androidx.test.core.app.ApplicationProvider;
 import androidx.test.filters.LargeTest;
 import de.blau.android.JavaResources;
 import de.blau.android.MockTileServer;
+import de.blau.android.layer.tiles.util.MapTileProviderCallback;
 import de.blau.android.resources.TileLayerDatabase;
 import de.blau.android.resources.TileLayerSource;
 import de.blau.android.resources.TileLayerSource.Category;
 import de.blau.android.resources.TileLayerSource.Provider;
 import de.blau.android.util.FileUtil;
-import de.blau.android.views.util.MapTileProviderCallback;
 import okhttp3.mockwebserver.MockWebServer;
 
 @RunWith(RobolectricTestRunner.class)
